refactor(projects): split project media image viewer
